+++ src/event_loop.rs
@@ -13,6 +13,7 @@
 //! [event_loop_proxy]: crate::event_loop::EventLoopProxy
 //! [send_event]: crate::event_loop::EventLoopProxy::send_event
 use instant::Instant;
+use raw_window_handle::{HasRawDisplayHandle, RawDisplayHandle};
 use std::{error, fmt, ops::Deref};
 
 use crate::{event::Event, monitor::MonitorHandle, platform_impl};
@@ -209,6 +210,13 @@ impl<T> EventLoopWindowTarget<T> {
   #[inline]
   pub fn primary_monitor(&self) -> Option<MonitorHandle> {
     self.p.primary_monitor()
+  }
+}
+
+unsafe impl<T> HasRawDisplayHandle for EventLoopWindowTarget<T> {
+  /// Returns a [`raw_window_handle::RawDisplayHandle`] for the event loop.
+  fn raw_display_handle(&self) -> RawDisplayHandle {
+    self.p.raw_display_handle()
   }
 }