+++ src/platform_impl/linux/x11/mod.rs
@@ -1,0 +1,4 @@
+pub mod ffi;
+pub mod xdisplay;
+
+pub use xdisplay::XConnection;